+++ xmtp_id/src/associations/builder.rs
@@ -1,3 +1,7 @@
+//! Builders for creating a [`SignatureRequest`] with a [`PendingIdentityAction`] for an external SDK/Library, which can then be
+//! resolved into an [`IdentityUpdate`]. An [`IdentityUpdate`] may be used for updating the state
+//! of an XMTP ID according to [XIP-46](https://github.com/xmtp/XIPs/pull/53)
+
 use std::collections::{HashMap, HashSet};
 
 use crate::utils::now_ns;
